BrainDead_2k26/BrainDead-Solution_PS2/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
einops>=0.6.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.2.0

# Medical Imaging
pydicom>=2.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training Utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Evaluation
nltk>=3.8.0

# Development
jupyter>=1.0.0
ipykernel>=6.23.0